requirements.txt
Django
django-cors-headers
djangorestframework
gunicorn
uvicorn
requests
python-dotenv
mysqlclient
django-filter
django-allauth 